get rid of Coins, use FT instead
