--- src/modules/coin/data_provider/mod.rs
@@ -1,8 +1,0 @@
-mod balance;
-mod history;
-mod metadata;
-mod models;
-
-pub(crate) use balance::{get_coin_balances, get_coin_balances_by_contract, get_near_balance};
-pub(crate) use history::{get_coin_history, get_near_history};
-pub(crate) use metadata::{get_ft_contract_metadata, get_near_metadata};
