+++ src/modules/ft/data_provider/metadata.rs
@@ -1,14 +1,14 @@
 use borsh::{BorshDeserialize, BorshSerialize};
 use serde::{Deserialize, Serialize};
 
-use crate::modules::coin;
+use crate::modules::ft;
 use crate::{rpc_helpers, types};
 
-pub(crate) async fn get_ft_contract_metadata(
+pub(crate) async fn get_ft_metadata(
     rpc_client: &near_jsonrpc_client::JsonRpcClient,
     contract_id: near_primitives::types::AccountId,
     block_height: u64,
-) -> crate::Result<coin::schemas::FtContractMetadata> {
+) -> crate::Result<ft::schemas::FtContractMetadata> {
     let request = rpc_helpers::get_function_call_request(
         block_height,
         contract_id.clone(),
@@ -19,7 +19,7 @@ pub(crate) async fn get_ft_contract_metadata(
         rpc_helpers::wrapped_call(rpc_client, request, block_height, &contract_id).await?;
 
     let metadata = serde_json::from_slice::<FtMetadata>(&response.result)?;
-    Ok(coin::schemas::FtContractMetadata {
+    Ok(ft::schemas::FtContractMetadata {
         spec: metadata.spec,
         name: metadata.name,
         symbol: metadata.symbol,
@@ -28,16 +28,6 @@ pub(crate) async fn get_ft_contract_metadata(
         reference: metadata.reference,
         reference_hash: types::vector::base64_to_string(&metadata.reference_hash)?,
     })
-}
-
-pub(crate) fn get_near_metadata() -> coin::schemas::CoinMetadata {
-    coin::schemas::CoinMetadata {
-        name: "NEAR blockchain native token".to_string(),
-        symbol: "NEAR".to_string(),
-        // TODO PHASE 2 re-check the icon. It's the best I can find
-        icon: Some("https://raw.githubusercontent.com/near/near-wallet/7ef3c824404282b76b36da2dff4f3e593e7f928d/packages/frontend/src/images/near.svg".to_string()),
-        decimals: 24,
-    }
 }
 
 // Taken from https://github.com/near/near-sdk-rs/blob/master/near-contract-standards/src/fungible_token/metadata.rs
@@ -52,9 +42,9 @@ pub struct FtMetadata {
     pub decimals: u8,
 }
 
-impl From<coin::schemas::FtContractMetadata> for coin::schemas::CoinMetadata {
-    fn from(metadata: coin::schemas::FtContractMetadata) -> Self {
-        coin::schemas::CoinMetadata {
+impl From<ft::schemas::FtContractMetadata> for ft::schemas::Metadata {
+    fn from(metadata: ft::schemas::FtContractMetadata) -> Self {
+        ft::schemas::Metadata {
             name: metadata.name,
             symbol: metadata.symbol,
             icon: metadata.icon,
@@ -75,7 +65,7 @@ mod tests {
         let block = get_block();
         let contract = near_primitives::types::AccountId::from_str("usn").unwrap();
 
-        let metadata = get_ft_contract_metadata(&rpc_client, contract, block.height).await;
+        let metadata = get_ft_metadata(&rpc_client, contract, block.height).await;
         insta::assert_debug_snapshot!(metadata);
     }
 
@@ -85,7 +75,7 @@ mod tests {
         let block = get_block();
         let contract = near_primitives::types::AccountId::from_str("olga.near").unwrap();
 
-        let metadata = get_ft_contract_metadata(&rpc_client, contract, block.height).await;
+        let metadata = get_ft_metadata(&rpc_client, contract, block.height).await;
         insta::assert_debug_snapshot!(metadata);
     }
 
@@ -95,7 +85,7 @@ mod tests {
         let block = get_block();
         let contract = near_primitives::types::AccountId::from_str("comic.paras.near").unwrap();
 
-        let metadata = get_ft_contract_metadata(&rpc_client, contract, block.height).await;
+        let metadata = get_ft_metadata(&rpc_client, contract, block.height).await;
         insta::assert_debug_snapshot!(metadata);
     }
 }
