+++ src/modules/ft/schemas.rs
@@ -5,7 +5,6 @@ use crate::types;
 
 // *** Requests ***
 
-// move to coins
 #[derive(
     Validate, Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, Apiv2Schema,
 )]
@@ -34,7 +33,6 @@ pub struct HistoryRequest {
     pub contract_account_id: types::AccountId,
 }
 
-// duplicate in each folder
 #[derive(
     Validate, Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, Apiv2Schema,
 )]
@@ -46,25 +44,26 @@ pub struct ContractMetadataRequest {
 // *** Responses ***
 
 #[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, Apiv2Schema)]
-pub struct NearBalanceResponse {
-    /// Sum of staked and nonstaked balances
-    pub balance: types::U128,
-    pub metadata: CoinMetadata,
+pub struct FtBalanceByContractResponse {
+    pub balance: FtBalance,
     pub block_timestamp_nanos: types::U64,
     pub block_height: types::U64,
 }
 
-/// This response gives the information about all the available balances for the user.
-/// The answer gives the list of NEAR, FT balances, could be used for Multi Tokens.
-/// For MTs and other standards, balances could have multiple entries for one contract.
 #[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, Apiv2Schema)]
-pub struct CoinBalancesResponse {
-    pub balances: Vec<Coin>,
+pub struct FtBalancesResponse {
+    pub balances: Vec<FtBalance>,
     pub block_timestamp_nanos: types::U64,
     pub block_height: types::U64,
 }
 
-/// This response provides the coin history (NEAR or by contract).
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, Apiv2Schema)]
+pub struct FtBalance {
+    pub amount: types::U128,
+    pub contract_account_id: types::AccountId,
+    pub metadata: Metadata,
+}
+
 #[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, Apiv2Schema)]
 pub struct HistoryResponse {
     pub history: Vec<HistoryItem>,
@@ -81,45 +80,25 @@ pub struct FtContractMetadataResponse {
 
 // ---
 
-/// This type describes general coin information.
-/// It is used for NEAR, FT balances, could be used for Multi Tokens.
-///
-/// For MTs and other standards, we could have multiple coins for one contract.
-/// For NEAR and FTs, coin_metadata contains general metadata (the only available option, though).
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, Apiv2Schema)]
-pub struct Coin {
-    /// "nearprotocol" for NEAR, "nep141" for FT
-    pub standard: String,
-    pub balance: types::U128,
-    /// null for NEAR, not null otherwise
-    pub contract_account_id: Option<types::AccountId>,
-    pub metadata: CoinMetadata,
-    // TODO PHASE 1 (idea) I think it would be great to add here the info about last update moment. Timestamp, later also index
-    // I'm already doing it at NftCount
-}
-
-/// This type describes the history of coin movements for the given user.
-/// Coins could be NEAR, FT, it could be also later used for Multi Tokens.
+/// This type describes the history of the operations (NEAR, FT) for the given user.
 #[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, Apiv2Schema)]
 pub struct HistoryItem {
     // TODO PHASE 2 add index here
     // pub index: types::U128,
-    // TODO PHASE 1 (idea) do we want to add here tx_hash/receipt_id? We may want to add it at many places
     pub involved_account_id: Option<types::AccountId>,
     pub delta_balance: types::I128,
     pub balance: types::U128,
     pub cause: String,
     pub status: String,
-    pub coin_metadata: CoinMetadata,
+    pub metadata: Metadata,
     pub block_timestamp_nanos: types::U64,
     // TODO PHASE 2 add this when we have all the data in the same DB. Now we can't join with blocks
     // pub block_height: types::U64,
 }
 
 /// This type describes general Metadata info, collecting the most important fields from different standards in the one format.
-/// `decimals` may contain `0` if it's not applicable (e.g. if it's general MT metadata)
 #[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, Apiv2Schema)]
-pub struct CoinMetadata {
+pub struct Metadata {
     pub name: String,
     pub symbol: String,
     pub icon: Option<String>,
