+++ src/modules/mod.rs
@@ -1,6 +1,7 @@
 use crate::{db_helpers, errors, types};
 
-pub(crate) mod coin;
+pub(crate) mod ft;
+pub(crate) mod native;
 pub(crate) mod nft;
 
 pub(crate) async fn check_account_exists(
