+++ src/modules/native/data_provider/mod.rs
@@ -1,0 +1,8 @@
+mod balance;
+mod history;
+mod metadata;
+mod models;
+
+pub(crate) use balance::get_near_balance;
+pub(crate) use history::get_near_history;
+pub(crate) use metadata::get_near_metadata;
