+++ src/modules/native/data_provider/models.rs
@@ -16,15 +16,3 @@ pub(crate) struct NearHistoryInfo {
     pub block_timestamp_nanos: BigDecimal,
     // pub block_height: super::types::U64,
 }
-
-#[derive(sqlx::FromRow)]
-pub(crate) struct CoinHistoryInfo {
-    // TODO PHASE 2 add symbol
-    // pub block_height: BigDecimal,
-    pub block_timestamp: BigDecimal,
-    pub amount: BigDecimal,
-    pub cause: String,
-    pub status: String,
-    pub old_owner_id: String,
-    pub new_owner_id: String,
-}
